fix: rename controller folder to plural
