+++ server/src/routes/user.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import UserController from '../controller/UserController';
+import UserController from '../controllers/UserController';
 
 const router = Router();
